requirements.txt
requests>=2,<3
python-dotenv<0.22
msal==1.30.0
